chore(flece:worker): move to sanitize namespace
