+++ merino-fleece/merino_fleece/sanitize/__init__.py
@@ -1,0 +1,1 @@
+"""Search term sanitization: models, sanitizer, emitter, and the queue worker."""
