+++ merino-fleece/merino_fleece/sanitize/worker/__init__.py
@@ -1,0 +1,1 @@
+"""Pub/Sub queue worker that consumes and sanitizes search term submissions."""
