+++ merino-fleece/tests/unit/sanitize/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for merino_fleece.sanitize."""
