+++ merino-fleece/tests/unit/sanitize/worker/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for merino_fleece.sanitize.worker."""
