+++ merino-fleece/tests/unit/sanitize/worker/test_worker.py
@@ -1,4 +1,4 @@
-"""Unit tests for merino_fleece.worker.worker."""
+"""Unit tests for merino_fleece.sanitize.worker.worker."""
 
 from typing import Any
 from unittest.mock import MagicMock
@@ -8,8 +8,8 @@ from pytest_mock import MockerFixture
 
 from merino_common.models.suggest_logging import SuggestRequestParams, SearchTermsSubmission
 
-from merino_fleece.models import SanitizedSuggestRequest
-from merino_fleece.worker.worker import FleeceQueueWorker, callback
+from merino_fleece.sanitize.models import SanitizedSuggestRequest
+from merino_fleece.sanitize.worker.worker import FleeceQueueWorker, callback
 
 
 def _log_entry(**overrides: Any) -> SuggestRequestParams:
@@ -30,7 +30,7 @@ def _log_entry(**overrides: Any) -> SuggestRequestParams:
 
 def test_callback_sanitizes_emits_and_acks(mocker: MockerFixture) -> None:
     """The callback sanitizes each term, emits it, and acks the message."""
-    emit = mocker.patch("merino_fleece.worker.worker.emit_sanitized_query")
+    emit = mocker.patch("merino_fleece.sanitize.worker.worker.emit_sanitized_query")
     submission = SearchTermsSubmission(search_terms=[_log_entry(), _log_entry()])
     message = MagicMock()
     message.data = submission.model_dump_json().encode("utf-8")
@@ -48,7 +48,7 @@ def test_callback_sanitizes_emits_and_acks(mocker: MockerFixture) -> None:
 
 def test_callback_empty_batch_still_acks(mocker: MockerFixture) -> None:
     """A submission with no search terms emits nothing but still acks."""
-    emit = mocker.patch("merino_fleece.worker.worker.emit_sanitized_query")
+    emit = mocker.patch("merino_fleece.sanitize.worker.worker.emit_sanitized_query")
     submission = SearchTermsSubmission(search_terms=[])
     message = MagicMock()
     message.data = submission.model_dump_json().encode("utf-8")
@@ -63,11 +63,11 @@ def test_callback_drops_invalid_message(
     mocker: MockerFixture, caplog: pytest.LogCaptureFixture
 ) -> None:
     """A message that fails validation is logged, dropped, and still acked."""
-    emit = mocker.patch("merino_fleece.worker.worker.emit_sanitized_query")
+    emit = mocker.patch("merino_fleece.sanitize.worker.worker.emit_sanitized_query")
     message = MagicMock()
     message.data = b'{"not": "a valid submission"}'
 
-    with caplog.at_level("ERROR", logger="merino_fleece.worker.worker"):
+    with caplog.at_level("ERROR", logger="merino_fleece.sanitize.worker.worker"):
         callback(message)
 
     emit.assert_not_called()
@@ -78,7 +78,7 @@ def test_callback_drops_invalid_message(
 @pytest.fixture
 def subscriber_mock(mocker: MockerFixture) -> MagicMock:
     """Patch the Pub/Sub SubscriberClient and return the client instance mock."""
-    client_cls = mocker.patch("merino_fleece.worker.worker.pubsub_v1.SubscriberClient")
+    client_cls = mocker.patch("merino_fleece.sanitize.worker.worker.pubsub_v1.SubscriberClient")
     instance: MagicMock = client_cls.return_value
     return instance
 
@@ -123,7 +123,7 @@ def test_stop_prevents_reconnect_after_error(
     subscriber_mock: MagicMock, mocker: MockerFixture
 ) -> None:
     """When stop() runs mid-stream, the resulting error does not trigger a reconnect."""
-    sleep = mocker.patch("merino_fleece.worker.worker.time.sleep")
+    sleep = mocker.patch("merino_fleece.sanitize.worker.worker.time.sleep")
     worker = FleeceQueueWorker("my-subscription")
     future = subscriber_mock.subscribe.return_value
 
@@ -145,14 +145,14 @@ def test_restarts_after_stream_error(
     subscriber_mock: MagicMock, mocker: MockerFixture, caplog: pytest.LogCaptureFixture
 ) -> None:
     """A streaming error is logged, the future cancelled, and the stream restarted."""
-    sleep = mocker.patch("merino_fleece.worker.worker.time.sleep")
+    sleep = mocker.patch("merino_fleece.sanitize.worker.worker.time.sleep")
     subscriber_mock.closed = True  # restart() only proceeds on a closed client
     future = subscriber_mock.subscribe.return_value
     # First stream dies; the restarted stream blocks then returns cleanly.
     future.result.side_effect = [RuntimeError("stream died"), None]
     worker = FleeceQueueWorker("my-subscription", restart_backoff=5)
 
-    with caplog.at_level("ERROR", logger="merino_fleece.worker.worker"):
+    with caplog.at_level("ERROR", logger="merino_fleece.sanitize.worker.worker"):
         worker.start()
 
     assert "Encountered exception during message streaming" in caplog.text
@@ -171,7 +171,7 @@ def test_restart_ignored_while_subscriber_open(
     worker = FleeceQueueWorker("my-subscription")
     subscriber_mock.subscribe.reset_mock()
 
-    with caplog.at_level("WARNING", logger="merino_fleece.worker.worker"):
+    with caplog.at_level("WARNING", logger="merino_fleece.sanitize.worker.worker"):
         worker.restart()
 
     subscriber_mock.subscribe.assert_not_called()
